feat: Memory-less convener agent and split up
